Continuação da implementação da pág Inclusão de pedidos

+++ LojasNet.UI.Web/Models/Item.cs
@@ -7,7 +7,7 @@
             public string ProdutoId { get; set; }
             public string Nome { get; set; }
             public int Quantidade { get; set; }
-            public int Valor { get; set; }
+            public decimal Valor { get; set; }
             public decimal Total
             {
                 get { return Valor * Quantidade; }
